renamed hex_ai/training_utils_legacy.py -> hex_ai/training_orchestration.py and updated calls.
